componentalized
